Update feature engineering script

+++ scripts/preprocess_data.py
@@ -54,40 +54,43 @@ def main():
     logger.info(f"Files to be preprocessed are: {raw_file_list}")
 
     out_file_dir = os.path.join(project_dir, "data", "interim")
-    df_list = []
-    for filename in raw_file_list:
-        raw_file = os.path.join(raw_file_dir, filename)
-        logger.info(f"Reading {raw_file}...")
-        df = pd.read_csv(raw_file)
-        logger.info(f"{raw_file} read.")
-        logger.info(f"Preprocessing {raw_file}")
-        df = prep.transform_column_names_to_lowercase(df)
-
-        time_columns = [
-            "adddate",
-            "resolutiondate",
-            "serviceduedate",
-            "serviceorderdate",
-            "inspectiondate",
-        ]
-        df = prep.convert_columns_to_datetime(df, time_columns)
-
-        df = prep.create_days_to_resolve_field(df)
-        df = prep.process_ward_field(df)
-        out_file_path = os.path.join(out_file_dir, filename)
-        logger.info(
-            f"Preprocessing of {raw_file} complete. Outputting data to {out_file_path}..."
-        )
-        df.to_csv(out_file_path, index=False)
-        logger.info(f"File successfully output to {out_file_path}.")
-        df_list.append(df)
-    logging.info("Stacking dataframes...")
-    df_all_years = pd.concat(df_list)
-    logging.info("Dataframes successfully stacked.")
     out_file_path = os.path.join(out_file_dir, "dc_311_preprocessed_data.csv")
-    logger.info(f"Saving stacked dataframe as CSV to {out_file_path}.")
-    df_all_years.to_csv(out_file_path, index=False)
-    logger.info("CSV successfully saved.")
+    if os.path.exists(out_file_path):
+        logger.info(f"{out_file_path} exists! Bypassing preprocessing.")
+    else:
+        df_list = []
+        for filename in raw_file_list:
+            raw_file = os.path.join(raw_file_dir, filename)
+            logger.info(f"Reading {raw_file}...")
+            df = pd.read_csv(raw_file)
+            logger.info(f"{raw_file} read.")
+            logger.info(f"Preprocessing {raw_file}")
+            df = prep.transform_column_names_to_lowercase(df)
+
+            time_columns = [
+                "adddate",
+                "resolutiondate",
+                "serviceduedate",
+                "serviceorderdate",
+                "inspectiondate",
+            ]
+            df = prep.convert_columns_to_datetime(df, time_columns)
+
+            df = prep.create_days_to_resolve_field(df)
+            df = prep.process_ward_field(df)
+            out_file_path = os.path.join(out_file_dir, filename)
+            logger.info(
+                f"Preprocessing of {raw_file} complete. Outputting data to {out_file_path}..."
+            )
+            df.to_csv(out_file_path, index=False)
+            logger.info(f"File successfully output to {out_file_path}.")
+            df_list.append(df)
+        logging.info("Stacking dataframes...")
+        df_all_years = pd.concat(df_list)
+        logging.info("Dataframes successfully stacked.")
+        logger.info(f"Saving stacked dataframe as CSV to {out_file_path}.")
+        df_all_years.to_csv(out_file_path, index=False)
+        logger.info("CSV successfully saved.")
 
 
 if __name__ == "__main__":
